fix(audit): 全量129项P0-P3审查修复 + 四轮交叉验证通过(R1-R4)
🔴 P0 (40项): SQL注入x2, 支付回调放行, IDOR防护x3, 魔数校验x3,
  逻辑删除11表V80, FK约束9条V81, 评价审核状态机, 课程审核API,
  分页偏移, 注册开关, Banner URL, 难度类型, UX骨架屏x4

🟠 P1 (41项): @AuditedLog 17处, XSS全覆盖, 日志注入防护,
  @Lazy构造器注入, 多角色JWT, 竖屏适配, 学期切换UI, 自动重试,
  视频水印, 包位置修正, MyReviewController独立

🟡 P2 (38项): 搜索优化, 内存泄露, aria-label动态更新, console日志,
  硬编码密码环境变量化, @TableField补齐, 权限修正, 图表错误,
  注释修正, 键盘导航, 侧边栏进度遮罩

⚪ P3 (10项): import清理, 过时注释更新, v-html安全确认,
  catch静默修复, 路由meta补齐, request.js常量提取,
  数据字典6处遗漏补齐, 注释头统一12文件

质量门禁: mvn compile 0 ERROR | npm build SUCCESS
  | precheck 13/13 | R1-R4 全部27项PASS

+++ micro-course-api/src/main/java/com/microcourse/service/ExerciseService.java
@@ -15,7 +15,7 @@ public interface ExerciseService {
 
     void delete(Long id);
 
-    PageResult<ExerciseVO> page(Integer courseId, Integer chapterId, Integer page, Integer size);
+    PageResult<ExerciseVO> page(Long courseId, Long chapterId, Integer page, Integer size);
 
     ExerciseVO getById(Long id);
